requirements.txt
numpy==1.26.4
Pillow==10.4.0
Flask==3.0.3
ortools==9.10.4067
